// src/shared/repositories/types/PaginationTypes.ts
/**
 * Copyright 2024 Brad Anderson
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 * @author Brad Anderson <BradA1878@pm.me>
 * @repository https://github.com/BradA1878/model-exchange-framework
 * @documentation https://brada1878.github.io/model-exchange-framework/
 */

/**
 * Pagination options for list queries.
 */
export interface PaginationOptions {
    /** Number of items to return (default: 20) */
    limit?: number;

    /** Number of items to skip (default: 0) */
    offset?: number;

    /** Sort field */
    sortBy?: string;

    /** Sort direction */
    sortOrder?: 'asc' | 'desc';

    /** Cursor for cursor-based pagination (alternative to offset) */
    cursor?: string;
}

/**
 * Paginated result wrapper.
 */
export interface PaginatedResult<T> {
    /** The items for this page */
    items: T[];

    /** Total count of items (if countable) */
    total: number;

    /** Whether there are more items */
    hasMore: boolean;

    /** Cursor for next page (if using cursor pagination) */
    nextCursor?: string;

    /** Current page info */
    pagination: {
        limit: number;
        offset: number;
        page: number;
        totalPages: number;
    };
}